pyproject.toml
[project]
name = "veripulse"
version = "0.1.0"
description = "Automated news aggregation, analysis, generation, and social media dissemination for Philippine news"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Veripulse Team"}
]
keywords = ["news", "social-media", "automation", "nlp", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "feedparser>=6.0.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "tweepy>=4.14.0",
    "facebook-sdk>=3.1.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "python-dateutil>=2.8.0",
    "rich>=13.0.0",
    "newspaper3k>=0.2.0",
    "lxml_html_clean>=0.2.0",
    "tenacity>=8.0.0",
    "apscheduler>=3.10.0",
    "loguru>=0.7.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
embeddings = [
    "sentence-transformers>=2.5.0",
    "torch>=2.0.0",
]

[project.scripts]
veripulse = "veripulse.cli.main:app"

[project.urls]
Homepage = "https://github.com/veripulse/veripulse"
Repository = "https://github.com/veripulse/veripulse"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["veripulse"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
